+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This release extends support to OneView Rest API version 2000 (OneView v5.40) fo
 
 #### Features supported
 - Connection Template
+- Enclosure
+- Enclosure Group
 - Ethernet Network
 - FC Network
 - FCOE Network
@@ -13,7 +15,10 @@ This release extends support to OneView Rest API version 2000 (OneView v5.40) fo
 - Hypervisor Manager
 - Network Set
 - OS Deployment Plan
+- Scope
 - Server Certificate
+- Server Hardware
+- Server Hardware Type
 - Server Profile
 - Server Profile Template
 - Storage Pool